rename batch data
