+++ lambdas/sftp-letters/src/domain/synthetic-batch-data.ts
